Move the contact support message into PurchaseMeta component
