+++ client/me/purchases/manage-purchase/style.scss
@@ -108,17 +108,26 @@
 	}
 }
 
-.manage-purchase__meta {
-	display: flex;
-	list-style: none;
-	margin: 0 -24px -16px;
-	overflow: auto;
+.manage-purchase__meta,
+.manage-purchase__contact-support {
+	margin: 0 -24px;
 	padding: 0 24px;
 	border-top: 1px solid lighten( $gray, 30% );
 
 	@include breakpoint( "<660px" ) {
 		margin: 0 -16px;
 		padding: 0 16px;
+	}
+}
+
+.manage-purchase__meta {
+	display: flex;
+	margin-bottom: -16px;
+	list-style: none;
+	overflow: auto;
+
+	@include breakpoint( "<660px" ) {
+		margin-bottom: 0;
 		flex-direction: column;
 	}
